# updated-swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Dataverse Custom Connector API",
    "version": "v1"
  },
  "paths": {
    "/api/Dataverse/tables": {
      "get": {
        "tags": [
          "Dataverse"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DataverseTable"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DataverseTable"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DataverseTable"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/Dataverse/tables/{tableName}": {
      "get": {
        "tags": [
          "Dataverse"
        ],
        "parameters": [
          {
            "name": "tableName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {}
              },
              "application/json": {
                "schema": {}
              },
              "text/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/Dataverse/tables/{tableName}/columns": {
      "get": {
        "tags": [
          "Dataverse"
        ],
        "parameters": [
          {
            "name": "tableName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {}
              },
              "application/json": {
                "schema": {}
              },
              "text/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/Dataverse/health": {
      "get": {
        "tags": [
          "Dataverse"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {}
              },
              "application/json": {
                "schema": {}
              },
              "text/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/Dataverse/debug/config": {
      "get": {
        "tags": [
          "Dataverse"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {}
              },
              "application/json": {
                "schema": {}
              },
              "text/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "tags": [
          "PowerPlatformConnector"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "DataverseTable": {
        "type": "object",
        "properties": {
          "@odata.type": {
            "type": "string",
            "nullable": true
          },
          "MetadataId": {
            "type": "string",
            "nullable": true
          },
          "LogicalName": {
            "type": "string",
            "nullable": true
          },
          "DisplayName": {
            "$ref": "#/components/schemas/DisplayNameValue"
          },
          "Description": {
            "$ref": "#/components/schemas/DisplayNameValue"
          },
          "SchemaName": {
            "type": "string",
            "nullable": true
          },
          "EntitySetName": {
            "type": "string",
            "nullable": true
          },
          "IsCustomEntity": {
            "type": "boolean"
          },
          "OwnershipType": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DisplayNameValue": {
        "type": "object",
        "properties": {
          "UserLocalizedLabel": {
            "$ref": "#/components/schemas/LocalizedLabel"
          }
        },
        "additionalProperties": false
      },
      "LocalizedLabel": {
        "type": "object",
        "properties": {
          "Label": {
            "type": "string",
            "nullable": true
          },
          "LanguageCode": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      }
    }
  }
}